Create RegisterActivity
This code will create a simple register form with two fields: username and password. The registerButton will be clicked when the user clicks the "Register" button. If the username or password fields are empty, an error message will be displayed. Otherwise, a new user account will be created, and the user will be redirected to the login screen